+++ deps/icu-small/source/tools/toolutil/package.cpp
@@ -1,3 +1,5 @@
+// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 *
@@ -1292,7 +1294,7 @@ void Package::setItemCapacity(int32_t max)
     exit(U_MEMORY_ALLOCATION_ERROR);
   }
   if(items && itemCount>0) {
-    uprv_memcpy(newItems, items, itemCount*sizeof(items[0]));
+    uprv_memcpy(newItems, items, (size_t)itemCount*sizeof(items[0]));
   }
   itemMax = max;
   items = newItems;
